BackendProcesses/requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
asttokens==2.0.5
audioread==2.1.9
backcall==0.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cmudict==1.0.2
colorama==0.4.5
cycler==0.11.0
decorator==4.4.2
executing==0.8.3
Flask==2.1.3
Flask-RESTful==0.3.9
fonttools==4.34.4
idna==3.3
imageio==2.20.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
inflect==5.3.0
ipdb==0.13.9
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
librosa==0.9.0
llvmlite==0.38.1
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
moviepy==1.0.3
noisereduce==2.0.1
numba==0.55.2
numpy==1.22.4
packaging==21.3
pandas==1.1.4
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pooch==1.6.0
proglog==0.1.10
prompt-toolkit==3.0.30
pure-eval==0.2.2
PyAudio==0.2.12
pycparser==2.21
pycryptodome==3.15.0
pydub==0.25.1
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytools==2022.1.12
pytz==2022.1
PyYAML==6.0
requests==2.28.1
resampy==0.3.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
SoundFile==0.10.3.post1
sox==1.4.1
stack-data==0.3.0
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
toml==0.10.2
# torch==1.12.0+cu116
# torchaudio==0.12.0+cu116
# torchvision==0.13.0+cu116
tqdm==4.53.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.10
wcwidth==0.2.5
Werkzeug==2.1.2
youtube-dl==2021.12.17
zipp==3.8.1
